added tests for extractors
