+++ tests/unit/data/test_image_properties.py
@@ -2,7 +2,7 @@ import itk
 import pytest
 
 from pyradise.data.image import ImageProperties
-from tests.unit.helpers.image_helpers import get_sitk_image, get_itk_image
+from tests.conftest import get_sitk_image, get_itk_image
 
 itk_img_1 = get_itk_image(seed=0, low=0, high=101, meta="nii")
 sitk_img_1 = get_sitk_image(seed=1, low=0, high=101, meta="nii")
