+++ tests/unit/data/test_intensity_image.py
@@ -2,7 +2,7 @@ import pytest
 
 from pyradise.data.image import IntensityImage
 from pyradise.data.modality import Modality
-from tests.unit.helpers.image_helpers import get_itk_image, get_sitk_image
+from tests.conftest import get_itk_image, get_sitk_image
 
 itk_img_1 = get_itk_image(seed=0, low=0, high=101, meta="nii")
 sitk_img_1 = get_sitk_image(seed=1, low=0, high=101, meta="nii")
